requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.1.0
python-dotenv==1.0.1
Flask-Mail==0.9.1
reportlab==4.0.0
numpy==1.26.1
pandas==2.2.2
scikit-learn==1.4.2
python-dateutil==2.8.2

